I think I found a souloution for the caption problem
The sendMedia method does not send the photos or music with the caption the caption is always none. I need to test each ckeckpoint where the caption goes to see in which one it dies, after that I need to fix the problem in that function that the capiton dies on, I will do this by insted of passin the funciton to next function I will print out the data to be sent and see the result. Also I need to change the library that the bot_actionm module uses to send api requests to the one that I used in the tests api calls.