fixed the caption problem
There was no problem, the problem was with my tests (-_-). I needed to pass the caption to the params not the data. I also cut the fat of the whole program, it was to complex. from now on I will increse the complexity of the program according to it's scale