I fixed the caption tests
The test that were for caption were not working. I realized the problem was that in the expected_result functin I put the expected_result variable in outsiede of the loop so any time that I would change it it would be rewritten insted being made a new for evry iteration. I simply put it back in the loop and vala. I also updated the documentation in the bot_action module