seshat_interaction making has begun
I made a while loop and in it I make the request to the api for updates. The next step would have been to create some begavior based on the update that the program recives from the api, but the allowed_updates param for the getUpdates method is flawed I need to fix that and than I will continue the procces of making algorithms for the based on the updates that the bot is reciving.