made the function to find the number of updates
I realized that the tests for the getUPdates method were not at all adequate so I want to make more test so that I can make sure all of the functionality of the method works when I need to, first i nedd to see if the  response is ok or not that test is complete next I need to see how many updates are in the update list and if they are more than 3 we can do some comprehensvie test with that data. right now I'm in the stage of figuring out if there are enough updates or not and if there are not enough updates, the testing script will ask the tester to send a few updates. or the tester can choose no or do nothing and the value of this test will come back neagative and all the future tests that will depend on it , like the offset param test that will come after this , will fail with the reason above given. That is the vision for the getUpdates method test for now.