I fixed the 'not enough updates for testing' problem
I made another bot and named it test_user and this will substitude the lack of user update if there are such problems in our testing. there is no longer a need for the tester to get involved the whole proccess is automatic.The next phase is to write the test that require actual updates to to throughly test the method. Like the offset param test or the selected_media param test. in all of this tests I first use the tools that I've made to see if there are enough updates for this test if true I proceed if not I will call the recive test_updates function and update the update_response varale and than proceed with the tests. perhaps I need the tools that I've at all I can jsut call the function to recive the updates and weather or not there is enough updates , it wouldn't make a difference. this will require more thought