I made the test fot offset param
the offset param test requires there to be to sets update made before and after the middle part of the test. I want to solve this problem with the help of the uset input and that will, I know it is a very iellagant way to solve the issiue. but if it works it works. ufortunatly it doesnt work at the moment. I fix the issure late