The offset test works!
the bot would not aknoledge the updates that would come from another bot, so I needed to make it so that the updates would come from a chat_id that registered as a real user. the best way to do that would have been to use the telegram api with a telegram api token. And that would have made the whole process automatic. Unfortonatly the aquisition of telegram api token in iran is restricted I'm certain there are ways around it but for now I wanted to make the test work so made the tester manualy message the bot the updates. in the future I need to keep an eye out for ways to get a telegram api token so that I could make this process automatic