I automated the offset test
I realized that the bot will aknoledge the channel posts as updates regardless of weather a bot posted them or a user so I put the test bot and the real bot and the test_user_bot in the test_channel and made all of them admins. and using the recive update functint the test_user bot will send sevral messages to the test channel and than we can test the getUpdate method and it's params automaticly now.