I modified the create_db script and the update_group_members function
I realized that the reliance on timestamps was very redundant and over complacating so I forog that. I also wanted to automate the db creation proccess so I changed the crate_db script so that it would check if the db exists and if not create it and than I imported that script to the main script and used on top of the scrip.