made the remove_member endpoint
first of all I moved the group_member type class to the data file for simplicity sake. next I made the removeGroupMember in more or less the fasion of the upate_group_member expect insted of adding it deletes. I also made some tests to make the endpoiunt does what it needs to do. The tests require a little more scrutiny than this endpoint will be 100% complete