I made the get_file_path function in the bot action
I used the telegram_request function in the bot_actions module to call the getFile method of the teleegram api and than and than the functin stores the file_path that it gets to the db for further use now I need to make a method to delete the photos. than i made tests for this func. I than mede the tests so that it will check for both a succsesfull function call and a failure than I made the tests by mocking the telegram_request function .