made the set_caht_photo function ans massively improved error handling
the error handling of the various function specially that dealt with the db and the telegram api was abysmal so I added various error handling for the functions that dealt with those two thing, I finshed the delete_photo_path and than I completed the set_chat_photo useing the stor_photo_path and the delete_photo_path and usin ghte set_chat_phto method in telegram api I made tests for all of theses. and at last I made the endpoint changeGroupPhoto using the set_chat_photo, of course there is error handling at each level. the endpoint does not work appearantly the filr needs to be an input file so I will have to add a step to the store and delelete function to not onl save the file path that is on the server but to downlod to the db and than deleting it.