the set_chat_photo function problem has been fixed
now insted of sending the file path of the photo to the telegram api request now we upload the file and than give it to the telegram api as input file witch was the problem before. The problem now is the testing, the tests don't work at all. I'm relying on alot of mock date to do the tests, I will either need to improve the mocking proccess or just simply use real file_id and be done with it. I think the second option is better.