The redesign is over
I changed many things but the main thing is that I created a file that is soley host of the functions that manipulate the db and it is through these functions that the rest of the api must interact with the db. The other main thing that I changed is testing I made it so that it would rely more on mocking and also I organized them a bit better. Other than that the other changes are minimal and are not really worth mentioning